refactor and restructure code for easier understanding
